# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tole Community Management API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Tole Community Management System",
      "url": "https://github.com/"
    },
    "license": {
      "name": "UNLICENSED"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "This server — all versioned endpoints live under the /api prefix."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the `token` returned by `POST /auth/login` into the Authorize dialog. It is sent as `Authorization: Bearer <token>`. Query-string tokens are rejected by the server."
      }
    },
    "schemas": {
...
      "House": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65f1a2b3c4d5e6f7a8b9c0d2"
          },
          "houseNo": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20,
            "pattern": "^[A-Z0-9][A-Z0-9\\-/ ]*$",
            "example": "A-101",
            "description": "Stored uppercase and unique across the community, case-insensitively."
          },
          "section": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9 \\-]*$",
            "example": "Section 1",
            "description": "Used to target notices and polls to specific sections."
          },
          "floor": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30,
            "example": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "apartment",
              "house",
              "shop"
            ],
            "description": "Kind of dwelling."
          },
          "owner": {
            "type": "object",
            "description": "Populated resident. `null` when unassigned. A resident may own at most one house.",
            "properties": {
              "_id": {
                "type": "string",
                "example": "65f1a2b3c4d5e6f7a8b9c0d1"
              },
              "name": {
                "type": "string",
                "minLength": 2,
                "maxLength": 50,
                "pattern": "^[A-Za-z][A-Za-z ]*$",
                "example": "Ram Bahadur"
              },
              "username": {
                "type": "string",
                "example": "ram.bahadur"
              },
              "phone": {
                "type": "string",
                "pattern": "^(?:98[012456]|97[01456])\\d{7}$",
                "example": "9841234567",
                "description": "Optional. A 10-digit Nepal mobile number starting with 980, 981, 982, 984, 985, 986, 970, 971, 974, 975, or 976."
              },
              "email": {
                "type": "string",
                "format": "email",
                "maxLength": 100,
                "example": "ram.bahadur@example.com"
              },
              "specialization": {
                "type": "string",
                "nullable": true,
                "enum": [
                  "water",
                  "electric",
                  "lift",
                  "sanitation",
                  "security",
                  "general",
                  null
                ],
                "description": "Only populated in projections that select it."
              }
            },
            "nullable": true
          },
          "tenant": {
            "type": "object",
            "description": "Populated resident. `null` when unassigned. A resident may tenant at most one house.",
            "properties": {
              "_id": {
                "type": "string",
                "example": "65f1a2b3c4d5e6f7a8b9c0d1"
              },
              "name": {
                "type": "string",
                "minLength": 2,
                "maxLength": 50,
                "pattern": "^[A-Za-z][A-Za-z ]*$",
                "example": "Ram Bahadur"
              },
              "username": {
                "type": "string",
                "example": "ram.bahadur"
              },
              "phone": {
                "type": "string",
                "pattern": "^(?:98[012456]|97[01456])\\d{7}$",
                "example": "9841234567",
                "description": "Optional. A 10-digit Nepal mobile number starting with 980, 981, 982, 984, 985, 986, 970, 971, 974, 975, or 976."
              },
              "email": {
                "type": "string",
                "format": "email",
                "maxLength": 100,
                "example": "ram.bahadur@example.com"
              },
              "specialization": {
                "type": "string",
                "nullable": true,
                "enum": [
                  "water",
                  "electric",
                  "lift",
                  "sanitation",
                  "security",
                  "general",
                  null
                ],
                "description": "Only populated in projections that select it."
              }
            },
            "nullable": true
          },
          "monthlyDue": {
            "type": "number",
            "exclusiveMinimum": 0,
            "maximum": 10000,
            "multipleOf": 0.01,
            "example": 500,
            "description": "Required positive monthly charge in NPR used when generating dues. The default maximum is configurable with MAX_MONTHLY_DUE."
          },
          "isOccupied": {
            "type": "boolean",
            "example": true,
            "description": "Only occupied, non-archived houses get dues generated."
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ],
            "description": "`archived` houses are excluded from dues and complaint lists unless `history=true` is requested."
          },
          "address": {
            "type": "string",
            "maxLength": 200,
            "example": "Lalitpur, Ward 4"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateHouseRequest": {
        "type": "object",
        "required": [
          "houseNo",
          "section",
          "floor",
          "monthlyDue"
        ],
        "properties": {
          "houseNo": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20,
            "pattern": "^[A-Z0-9][A-Z0-9\\-/ ]*$",
            "example": "A-101"
          },
          "section": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9 \\-]*$",
            "example": "Section 1"
          },
          "floor": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30,
            "example": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "apartment",
              "house",
              "shop"
            ],
            "description": "Kind of dwelling."
          },
          "owner": {
            "type": "string",
            "description": "User ID of an active resident. Omit or pass an empty string for none.",
            "example": "65f1a2b3c4d5e6f7a8b9c0d1"
          },
          "tenant": {
            "type": "string",
            "description": "User ID of an active resident. Omit or pass an empty string for none.",
            "example": "65f1a2b3c4d5e6f7a8b9c0d3"
          },
          "monthlyDue": {
            "type": "number",
            "exclusiveMinimum": 0,
            "maximum": 10000,
            "multipleOf": 0.01,
            "example": 500
          },
          "isOccupied": {
            "type": "boolean",
            "default": true
          },
          "address": {
            "type": "string",
            "maxLength": 200,
            "example": "Lalitpur, Ward 4"
          }
        }
      },
      "UpdateHouseRequest": {
        "type": "object",
        "description": "Partial update. To unassign an owner or tenant send an empty string (`\"\"`) for that key — omitting the key leaves it untouched.",
        "properties": {
          "houseNo": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20,
            "pattern": "^[A-Z0-9][A-Z0-9\\-/ ]*$",
            "example": "A-102"
          },
          "section": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9 \\-]*$",
            "example": "Section 2"
          },
          "floor": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30,
            "example": 2
          },
          "type": {
            "type": "string",
            "enum": [
              "apartment",
              "house",
              "shop"
            ],
            "description": "Kind of dwelling."
          },
          "owner": {
            "type": "string",
            "example": ""
          },
          "tenant": {
            "type": "string",
            "example": ""
          },
          "monthlyDue": {
            "type": "number",
            "exclusiveMinimum": 0,
            "maximum": 10000,
            "multipleOf": 0.01,
            "example": 750
          },
          "isOccupied": {
            "type": "boolean",
            "example": true
          },
          "address": {
            "type": "string",
            "maxLength": 200,
            "example": "Lalitpur, Ward 4"
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "description": "A single failed field check, as produced by `express-validator`.",
        "properties": {
          "field": {
            "type": "string",
            "example": "email"
          },
          "message": {
            "type": "string",
            "example": "Invalid email format"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Access denied"
          },
          "code": {
            "type": "string",
            "example": "RATE_LIMITED",
            "description": "Machine-readable error code when provided."
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "errors"
        ],
        "description": "The `express-validator` failure envelope. Note the contrast with the export-filter error, whose `errors` array contains plain strings instead of `{ field, message }` objects.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "description": "Endpoints that return no resource, only a status message.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed."
          }
        }
      }
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "Page number, 1-based. **Omit this entirely to disable pagination** and receive the full result set — several frontend screens (house pickers, section pickers) rely on that behaviour.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1,
          "example": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Items per page. Clamped to the range 1-100. Has no effect unless `page` is also supplied.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20,
          "example": 20
        }
      },
...
      "IdPathParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "MongoDB ObjectId of the resource.",
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-fA-F]{24}$",
          "example": "65f1a2b3c4d5e6f7a8b9c0d1"
        }
      },
...
    },
    "responses": {
      "UnauthorizedResponse": {
        "description": "Authentication failed or was not supplied. Returned when the `Authorization` header is missing, the token is invalid or expired, the referenced user no longer exists, the account is deactivated, or a token was passed as a `token` / `access_token` **query parameter** (which this API rejects on purpose to avoid leaking credentials in logs).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "examples": {
          "missingToken": {
            "summary": "No Authorization header",
            "value": {
              "success": false,
              "message": "Not authorized"
            }
          },
          "queryToken": {
            "summary": "Token supplied in the query string",
            "value": {
              "success": false,
              "message": "Use the Authorization request header"
            }
          },
          "inactive": {
            "summary": "Deactivated account",
            "value": {
              "success": false,
              "message": "User account is inactive or not found"
            }
          },
          "invalid": {
            "summary": "Malformed or bad signature",
            "value": {
              "success": false,
              "message": "Token invalid"
            }
          }
        }
      },
      "ForbiddenResponse": {
        "description": "The caller is authenticated but not permitted. Covers role mismatches (`authorize`), record-level ownership checks (a resident touching another resident's due or complaint), and denied export attempts.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "examples": {
          "role": {
            "summary": "Wrong role for this endpoint",
            "value": {
              "success": false,
              "message": "Access denied"
            }
          },
          "export": {
            "summary": "Export not permitted for this staff member",
            "value": {
              "success": false,
              "message": "You are not allowed to export dues data."
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "The resource does not exist. Also returned when a file is requested that is missing on disk, and for malformed ObjectIds (`errorHandler` maps Mongoose `CastError` to 404).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ConflictResponse": {
        "description": "The request conflicts with the current state of the resource. Typical causes: a duplicate unique value, a payment or complaint that has already been processed by someone else, or a vote that has already been cast.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "examples": {
          "duplicate": {
            "summary": "Unique constraint violation",
            "value": {
              "success": false,
              "message": "House number already exists"
            }
          },
          "processed": {
            "summary": "Already handled by a concurrent request",
            "value": {
              "success": false,
              "message": "This payment was already processed. Refresh and try again."
            }
          },
          "alreadyVoted": {
            "summary": "Duplicate vote",
            "value": {
              "success": false,
              "message": "You have already voted or this poll is no longer available."
            }
          }
        }
      },
...
      "ValidationFailedResponse": {
        "description": "One or more fields failed `express-validator` checks. Any files uploaded alongside the request are deleted before responding.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        },
        "example": {
          "success": false,
          "message": "Validation failed",
          "errors": [
            {
              "field": "email",
              "message": "Invalid email format"
            },
            {
              "field": "password",
              "message": "Password must be at least 6 characters"
            }
          ]
        }
      },
...
      "ServerErrorResponse": {
        "description": "Unhandled server error. `message` is the raw error text, so treat it as diagnostic only.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/houses": {
      "get": {
        "operationId": "listHouses",
        "summary": "List houses",
        "description": "Returns houses ordered by section, then house number. `owner` and `tenant` are populated with `name`, `username`, `phone` and `email`.\n\n**Residents only ever see their own linked houses** — the server resolves the resident's `resident_houses` links (falling back to the legacy `house.owner`/`house.tenant` fields) and ignores any attempt to widen the result set. Admins and staff see everything.\n\n**Pagination is opt-in.** Send `page` to paginate; omit it to receive all 48 houses, which is what the frontend house pickers rely on.",
        "tags": [
          "Houses"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Matching houses with pagination metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 20
                    },
                    "total": {
                      "type": "integer",
                      "example": 48
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "pages": {
                      "type": "integer",
                      "example": 3
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/House"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      },
      "post": {
        "operationId": "createHouse",
        "summary": "Create a house",
        "description": "Registers a new house. `owner` and `tenant` must each reference an **active resident**; they cannot be the same person, and a resident may hold at most one house in each role (enforced by unique partial indexes).\n\nCreating the house also writes matching `resident_houses` link rows, which is what powers resident-scoped visibility across dues, complaints, notices and polls.",
        "tags": [
          "Houses"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "The house was created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/House"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailedResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "409": {
            "$ref": "#/components/responses/ConflictResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
    "/houses/{id}": {
      "put": {
        "operationId": "updateHouse",
        "summary": "Update a house",
        "description": "Partial update — only the keys present in the body are applied.\n\n**Unassigning:** send an empty string (`\"owner\": \"\"` or `\"tenant\": \"\"`) to clear that link. Omitting the key entirely leaves the existing link alone. Changing either one rewrites the corresponding `resident_houses` rows.",
        "tags": [
          "Houses"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The updated house.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/House"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailedResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/ConflictResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      },
      "delete": {
        "operationId": "archiveHouse",
        "summary": "Archive a house",
        "description": "**This does not delete anything.** The house is archived so that its dues, complaints and payment history stay intact for auditing:\n\n1. `isOccupied` is set to `false` and `owner`/`tenant` are cleared.\n2. Its `resident_houses` link rows are removed.\n\nThe `status` field becomes `archived`, which excludes the house from dues generation and from the default dues and complaints listings. Pass `history=true` to those list endpoints to include it again.",
        "tags": [
          "Houses"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The house was archived.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
...
  }
}